chore: move auth stuff to a specific auth file
